Critical fix: COBS decode trailing zero stripping caused data corruption
ROOT CAUSE: cobs_decode() stripped trailing zero bytes from decoded
output. When a COBS packet's CRC32 had 0x00 as its MSB (LE last byte),
the decode removed it, producing a 1-byte-shorter output. This caused
CRC mismatch for ~1/256 of all packets — deterministic, data-dependent.

Found via ASAN: "left shift of 136 by 24 places cannot be represented
in type 'int'" led to investigating CRC byte extraction, which led to
the COBS decode length mismatch.

Fixes:
- Remove trailing zero stripping from cobs_decode() (C)
- Cast uint8_t to uint32_t before << 24 in CRC extraction (UB fix)
- Per-packet backpressure ACK in WRITE and SELFUPDATE
- Fixed _recv_packet_sync: no partial frame stashing
- recv_response: reset deadline after READY skip
- ASAN firmware data test catches the bug offline

+++ agent/protocol.c
@@ -164,7 +164,8 @@ uint8_t proto_recv(uint8_t *data, uint32_t *len, uint32_t timeout_ms) {
     /* Verify CRC32 — read bytes individually to avoid unaligned access */
     uint32_t payload_len = raw_len - 4;
     volatile uint8_t *cp = &rx_raw[payload_len];
-    uint32_t expected_crc = cp[0] | (cp[1] << 8) | (cp[2] << 16) | (cp[3] << 24);
+    uint32_t expected_crc = (uint32_t)cp[0] | ((uint32_t)cp[1] << 8) |
+                           ((uint32_t)cp[2] << 16) | ((uint32_t)cp[3] << 24);
     uint32_t actual_crc = crc32(0, rx_raw, payload_len);
 
     /* Drain after CRC */
